==> _previous/clustering/albou558/can_2025/vote/03-6_issues.ipynb <==
# Library
library(ggplot2)
library(dplyr)

# Load Data and Clustering Results
dfIssues <- readRDS("../../../_SharedFolder_datagotchi_federal_2024/data/pilote/dataClean/datagotchi2025_canada_pilotClustering_20250310.rds")
dfUsedForClustering <- readRDS("../../../_PrivateFolder_datagotchi_federal_2025/data/clustering/can2025/03_pilot_2025.rds")

# Join both datasets to have cluster, issues and id

# Sélectionner les variables issues et politiques de dfIssues
dfIssues_selected <- dfIssues %>%
  select(id, starts_with("issue_"), dv_voteChoice, dv_attitudeLeftvsRight)

# Sélectionner les variables clusters de dfUsedForClustering
dfClusters_selected <- dfUsedForClustering %>%
  select(id, cluster_name, starts_with("cluster_"))

# Fusionner les deux jeux de données sur id
dfValidation <- dfClusters_selected %>%
  inner_join(dfIssues_selected, by = "id")

# Voir les variables de ce jeu fusionné
names(dfValidation)

generate_graph <- function(data, issue_variable, output_path) {
  
  # Convertir la variable en symbole pour dplyr et ggplot
  issue_sym <- sym(issue_variable)  # Changement ici

  # Calculer proportions et créer le graphique
  data %>%
    group_by(cluster_name, !!issue_sym) %>%
    summarise(nombre_de_repondants = n(), .groups = "drop") %>%
    group_by(cluster_name) %>%
    mutate(
      nombre_total_de_repondants = sum(nombre_de_repondants),
      proportion = nombre_de_repondants / nombre_total_de_repondants
    ) %>%
    ggplot(aes(x = !!issue_sym, y = proportion, fill = factor(cluster_name))) +
    geom_col() +
    facet_wrap(~ cluster_name) +
    labs(
      x = issue_variable,  # Correction pour utiliser le nom en string
      y = "Proportion",
      fill = "Cluster",
      title = paste("Distribution of", issue_variable, "by Cluster")
    ) +
    theme_minimal() -> graph
  
  # Sauvegarder le graphique
  ggsave(
    filename = paste0(output_path, "/validationCluster_", issue_variable, ".pdf"),
    plot = graph,
    width = 16,
    height = 12
  )
}


# Lister toutes les variables d'enjeux

issues_list <- colnames(dfValidation)[grepl("^issue_", colnames(dfValidation))]
issues_list <- c(issues_list, "dv_voteChoice", "dv_attitudeLeftvsRight")


# Lancer une boucle pour générer tous les graphiques

# Spécifier le dossier où enregistrer les graphiques
output_path <- "../../_PrivateFolder_datagotchi_federal_2025/graph/clustering/can_2025"

# Générer un graphique pour chaque enjeu sélectionné
lapply(issues_list, function(issue) {
  generate_graph(dfValidation, issue, output_path)
})
